requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
flatbuffers==23.5.26
fonttools==4.44.3
gast==0.4.0
gin==0.1.6
gin-config==0.5.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
joblib==1.3.2
keras==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
libscrc==1.8.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
memory-profiler==0.61.0
ml-dtypes==0.3.2
namex==0.0.7
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
objgraph==3.6.1
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.13
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==5.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
SciencePlots==2.1.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.24.0
termcolor==2.3.0
tf_keras==2.16.0
threadpoolctl==3.3.0
typing_extensions==4.5.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
